rename modules & add solid framework config (initial)
