+++ packages/builder-config/framework-solid/solid-config-styling.ts
@@ -1,0 +1,18 @@
+export type SolidStylingStrategy =
+  | SolidStyledComponentsConfig
+  | SolidInlineCssConfig;
+
+export type SolidStyledComponentsConfig = SolidOfficialStyledComponentsConfig;
+
+/**
+ * the official styled-components module by solidjs team
+ * > https://github.com/solidjs/solid-styled-components
+ */
+interface SolidOfficialStyledComponentsConfig {
+  type: "styled-components";
+  module: "solid-styled-components";
+}
+
+export interface SolidInlineCssConfig {
+  type: "inline-css";
+}
