+++ packages/builder-config/framework-solid/solid-output.ts
@@ -1,0 +1,3 @@
+import { ComponentOutput } from "../output";
+
+export interface SolidComponentOutput extends ComponentOutput {}
